Template for creating environment based on cloudinit images
- Add cloudinit-env.yaml fuel-devops template, which use
  cloudinit-meta-data.yaml and cloudinit-user-data.yaml to
  configure cloudinit image after start.
- cloudinit-user-data.yaml has iptables commands that allow access
  to port 22 only when all specified commands are completed.
  Tests should wait for port 22.

To use the template, a path to Ubuntu cloud image in qcow2 format
should be exported in the environment variable CLOUD_IMAGE_PATH.

No special preparation for the image is required, just download
and use. User 'vagrant' with sudo priveleges will be created
as specified in the cloudinit-user-data.yaml .

Note: IMAGE_PATH environment variable is not used in this template.

The template works with fuel-devops patch [1]

[1] https://review.openstack.org/#/c/338496/

Change-Id: I8771ff3adcd60b169787eee85b58918b7f61941e