Use 'backing_store' libvirt feature for create nodes
Patch [1] enables support for backing store volumes in fuel-devops.

Let's support this feature in the fuel-devops templates, which
dramatically increase speed of creation devops environments.

With this feature, an image that specified in 'group_volumes',
will be uploaded to the libvirt storage pool just once, and then
will be used for all other volumes (that relied on it) as a backing
storage without copying any data.

- add 'group_volumes' object to 'group', that contains list
  of volumes for use as backing volumes on nodes;
- add 'backing_store' options that contains the name of a volume
  from 'group_volumes'.
- desired volume size should be set with 'capacity' option in node
  objects. Thus, if original qcow2 image has 4Gb virtual size, but
  capacity is 50Gb, then resulting node volume will be 50Gb.

[1] https://review.openstack.org/#/c/333224/

Change-Id: I85d2655d88ee1bd4f4e02ccf110f8e5c95493446