Refactor fixtures that check required files and settings
Refactor those fixtures in a more general way, so they can be used in
any test, not only netchecker one.
Just set class variables required_files and required_settings to use them
In that way all required files and settings will be in the test class, and not
in the fixture itself.

Change-Id: I27eb98e46978205f4a4e9769e368d4720e2f8b55