[DO NOT REVIEW] Check k8s setup with enabled Calico network policies
Sometimes dnsmasq on node3 doesn't work correctly if Calico network policies
are enabled in kubernetes. It happens because dnsmasq pods are started before
Calico CNI binraries are copied to node3 (non-master node), so dnsmasq
is running in "offline" mode.

This patch is NOT going to be merged and created for temporary test purposes
only. If you are a reviewer of 'fuel-ccp-tests' - please just ignore it.

Change-Id: I206c13b1443572ac6db58313ad0cbe91125d7c1c